src/production/requirements.txt
pyyaml
polars
pyarrow
gcsfs
hawk-backtester
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp